refactor: payment params casting update

+++ core/eth/reader.go
@@ -867,7 +867,7 @@ func (t *Reader) GetGlobalSymbolsPerSecond(ctx context.Context) (uint64, error) 
 	return globalSymbolsPerSecond, nil
 }
 
-func (t *Reader) GetGlobalRatePeriodInterval(ctx context.Context) (uint32, error) {
+func (t *Reader) GetGlobalRatePeriodInterval(ctx context.Context) (uint64, error) {
 	if t.bindings.PaymentVault == nil {
 		return 0, errors.New("payment vault not deployed")
 	}
@@ -877,10 +877,10 @@ func (t *Reader) GetGlobalRatePeriodInterval(ctx context.Context) (uint32, error
 	if err != nil {
 		return 0, err
 	}
-	return uint32(globalRateBinInterval), nil
+	return globalRateBinInterval, nil
 }
 
-func (t *Reader) GetMinNumSymbols(ctx context.Context) (uint32, error) {
+func (t *Reader) GetMinNumSymbols(ctx context.Context) (uint64, error) {
 	if t.bindings.PaymentVault == nil {
 		return 0, errors.New("payment vault not deployed")
 	}
@@ -890,10 +890,10 @@ func (t *Reader) GetMinNumSymbols(ctx context.Context) (uint32, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint32(minNumSymbols), nil
+	return minNumSymbols, nil
 }
 
-func (t *Reader) GetPricePerSymbol(ctx context.Context) (uint32, error) {
+func (t *Reader) GetPricePerSymbol(ctx context.Context) (uint64, error) {
 	if t.bindings.PaymentVault == nil {
 		return 0, errors.New("payment vault not deployed")
 	}
@@ -903,10 +903,10 @@ func (t *Reader) GetPricePerSymbol(ctx context.Context) (uint32, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint32(pricePerSymbol), nil
+	return pricePerSymbol, nil
 }
 
-func (t *Reader) GetReservationWindow(ctx context.Context) (uint32, error) {
+func (t *Reader) GetReservationWindow(ctx context.Context) (uint64, error) {
 	if t.bindings.PaymentVault == nil {
 		return 0, errors.New("payment vault not deployed")
 	}
@@ -915,7 +915,7 @@ func (t *Reader) GetReservationWindow(ctx context.Context) (uint32, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint32(reservationWindow), nil
+	return reservationWindow, nil
 }
 
 func (t *Reader) GetOperatorSocket(ctx context.Context, operatorId core.OperatorID) (string, error) {
